# backend/docs/swagger.yaml
basePath: /
definitions:
  fiber.Map:
    additionalProperties: true
    type: object
  handler.LoginRequest:
    properties:
      password:
        description: 'required: true'
        example: password123
        type: string
      username:
        description: 'required: true'
        example: john_doe
        type: string
    type: object
...
  handler.RefreshRequest:
    properties:
      refresh_token:
        description: 'required: true'
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh...
        type: string
    type: object
  handler.RegisterRequest:
    properties:
      email:
        description: 'required: true'
        example: john@example.com
        type: string
      password:
        description: 'required: true'
        example: password123
        type: string
      role:
        example: owner
        type: string
      username:
        description: 'required: true'
        example: john_doe
        type: string
    type: object
...
  handler.TokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh...
        type: string
    type: object
  handler.User:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API for Order System with user authentication and restaurant management
  title: Order System API
  version: "1.0"
paths:
...
  /api/user/:
    delete:
      description: Delete the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Could not retrieve or delete user
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User
    get:
      description: Get all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.User'
            type: array
        "500":
          description: Could not retrieve users
          schema:
            type: string
      summary: Get all users
      tags:
      - User
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Login with username and password to get JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Invalid username or password
          schema:
            type: string
      summary: User login
      tags:
      - User
  /api/user/profile:
    get:
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.User'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
  /api/user/refresh:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokenResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            type: string
      summary: Refresh access token
      tags:
      - User
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid input
          schema:
            type: string
        "409":
          description: Username or email already taken
          schema:
            type: string
      summary: Register a new user
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"